move processing functions to own directory
